OMAP: CPU: CPUFREQ: CLOCK: VOLTAGE: LPDD: BOARD: PM:
Several fixes for stability to run an stable OPP frequency table with 10 - 12 OPPs. Prior to this commit we changed vp_timeout to u8 to allow boot OMAP with more than 7 OPPs. While browsing the mach-omap2 folder I found some mistakes from ealier commits and fixed these. In my testing phase with the new OPP table I faced random crashes/reboots which were related to voltage/sr calibration rule. Now trying the kernel with including these fixes. We may switch partly or completely to EMIF or SR_CLASS 1.5 if we can't get this to work.

Work left to do:

USB might be unstable related to usb performance boost mode - Not the boost is the problem but the voltage calibration, so whether we have to fix SR/switch to static voltages (but this may cause higher power consumption, possibly it also helps to set single OPPs to fixed, non-scalabe voltage) - or remove the variable usb performance mode and set a fixed frequency value (600mhz min/max).

In current kexec kernel we are facing an issue which causes the user cpufreq policy to be reset to kernel default in suspend/idle mode. This can be simply override by user setting an app dependant frequency lock. This issue also causes the @default_available rule for OPPs to be not effective. As we want to limit the default max frequency at 1.0GHZ we have to solve this problem or just define a new rule in omap2plus-cpufreq.c to prevent policy auto-reset.