Make generic policy rule for frequencies below 300mhz
The policy sacing from user space is broken in someway related to kexec. So we have to set a special policy rule to keep frequencies below 300mhz without the need of frequency lock from app side. So the user can set a preferred min freq while the max policy is untouched.