rpmsg: omx: initialize completion event in rpmsg_omx_open rpmsg: omx: initialize completion event in rpmsg_omx_open
An omx object is created when opening the rpmsg-omx device. However
the reply_arrived completion event is not initialized until the
connection message is sent across.

When the rpmsg_omx driver is removed (like in recovery), all pending
threads waiting on the reply_arrived are unblocked by a complete_all
call, however, if the rpmsg-omx device was only opened, such event is
not initialized yet, so the complete_all call will dereference a null
object causing a panic.

This patch moves the initialization of the reply_arrived completion
event to the open call, after the allocation of the omx object.