Revert "omap: serial: disable serial port irq during suspend"
This reverts commit 0af4d0c479b8ec6d9b8bc329b193a07331672fea.

This change does not work well together with Motorola's changes
to drivers/tty/serial/omap-serial.c and causes sporadic problems
with the serial ports resulting in the message "UART0: sw fifo
not empty" after resuming from sleep and the port then getting
stuck in a bad state until the next reboot.

This also causes the connected modem not to be accessible until
the device is rebooted.

Change-Id: I1e7a488d5607f3af8e113aace2a16556fd2c126a